+++ packages/editor/editor-core/examples/5-full-page-with-i18n.tsx
@@ -23,6 +23,7 @@ export default class ExampleEditor extends React.Component<Props, State> {
         messages={locales[locale]}
       >
         {FullPageExample({
+          allowHelpDialog: true,
           primaryToolbarComponents: (
             <WithEditorActions
               render={actions => (